# Performance optimization to parsing of floating point numbers